Thank you Colin! Transfer code to ZoweTreeProvider. Retained setProfile and Session to utilities because of RefreshTree
